fix: Commas fn accepts negative numbers

+++ src/common.rs
@@ -55,9 +55,9 @@ pub fn p(s: &str) -> String {
 
 // Adds commas to a string
 pub fn commas_from_string(n: &str) -> String {
-    let n = match n.parse::<u32>() {
+    let n = match n.parse::<f32>() {
         Ok(n) => n,
-        Err(_) => 0,
+        Err(_) => 0.0,
     };
 
     let num = NumberFormat::new();
